Implement internal logging mechanism.

+++ src/xv/sec_video_display.c
@@ -460,7 +460,7 @@ SECDisplayVideoSetPortAttribute (ScrnInfoPtr pScrn,
     if (attribute == _portAtom (PAA_OVERLAY))
     {
         pPort->overlay = value;
-        XDBG_DEBUG (MDA, "overlay(%d) \n", value);
+        XDBG_DEBUG (MDA, "overlay(%d) \n", (int) value);
         return Success;
     }
